+++ shared/util/common.lua
@@ -2,7 +2,6 @@
 
 
 require('__shared/util/sectorData')
-
 g_sectorTable = loadSectorTable()
 --print(g_sectorTable)
 
@@ -15,4 +14,8 @@ function changeSectorPoints(text,instance)
     for _,point in pairs(g_sectorTable[text]) do
         instance.points:add(point)
     end
+end
+
+function getSectorData(text)
+    return g_sectorTable[text]
 end