+++ ChangeLog.md
@@ -1,3 +1,4 @@
 # Changes per commit
 - 11/11/20 - Added US sector 1 Map vector Data and Flag data from CQ-L
 - 11/12/20 - Seperated the logic in various files and updated GUID for TDM
+- 11/13/20 - Added data points for all sectors